+++ backend/controllers/Partners.js
@@ -30,9 +30,11 @@ export const createPartner = async (req, res) => {
     const file = req.files.file;
     const fileSize = file.data.length;
     const ext = path.extname(file.name);
-    const fileName = file.md5 + ext;
+    const date = new Date();
+    const fileName = `${date.getHours()}${date.getMinutes()}${date.getSeconds()}${file.md5}` + ext;
     const urlImage = `${req.protocol}://${req.get("host")}/images/partners/${fileName}`;
     const allowedType = ['.png', '.jpg', '.jpeg'];
+
 
     if (!allowedType.includes(ext.toLowerCase())) return res.status(422).json({ msg: "Invalid Images" });
     if (fileSize > 5000000) return res.status(422).json({ msg: "Image must be less than 5 MB" });
@@ -63,14 +65,17 @@ export const updatePartner = async (req, res) => {
         const file = req.files.file;
         const fileSize = file.data.length;
         const ext = path.extname(file.name);
-        fileName = file.md5 + ext;
+        const date = new Date();
+        fileName = `${date.getHours()}${date.getMinutes()}${date.getSeconds()}${file.md5}` + ext;
         const allowedType = ['.png', '.jpg', '.jpeg'];
 
         if (!allowedType.includes(ext.toLowerCase())) return res.status(422).json({ msg: "Invalid Images" });
         if (fileSize > 5000000) return res.status(422).json({ msg: "Image must be less than 5 MB" });
 
         const filepath = `./public/images/partners/${partner.image}`;
-        fs.unlinkSync(filepath);
+        if (fs.existsSync(filepath)) {
+            fs.unlinkSync(filepath);
+        }
 
         file.mv(`./public/images/partners/${fileName}`, (err) => {
             if (err) return res.status(500).json({ msg: err.message });
@@ -101,7 +106,9 @@ export const deletePartner = async (req, res) => {
 
     try {
         const filepath = `./public/images/partners/${partner.image}`;
-        fs.unlinkSync(filepath);
+        if (fs.existsSync(filepath)) {
+            fs.unlinkSync(filepath);
+        }
         await Partner.destroy({
             where: {
                 id: req.params.id
